Also remove position style after going from affix-bottom to affix
The elemnt is relatively positioned when switches to affix-bottom. This
is done by calling $.offset, which sets both top and position: relative.

When switching back from affix-bottom, we also need to remove position:
relative.

Calling $.css('position', '') will remove the position part of the style
attribute so that whatever is defined elsewhere starts to be effective
again. This might clash with people who set position in the attribute,
but that also clashed before and isn't considered good practice anyways.